refactor(local-currency): extract CurrencyItem and debounce filtering
